move createQuery to separate file
